create terminal test helpers, move writeP there
